Fix connection change in multi-tenancy environment
This patch fix a problem when changing from the
provisioning network to the tenant one. When trying
to make the change the driver was trying to create
a 'Secondary' connection when there is no 'Primary'
anymore in the connection list. The method for create
port was updated. Now the check for if the next
connection creation will be 'Primary' or 'Secondary'
is performed after the previews connection removal
from the connection list.

Besides that, fix delete/create connection operations
and also make improvements debug/info logs.