Introduces developer mode on networking oneview
This patch fix the issue of launching a new OSA with the
networking_oneview installed. After the installation the
ml2 driver's synchronization mechanism delete networks
that are not related to the installed OSA. This is bad
because deletes networks that are under use by other
developers.

For solving the issue, we introduce the developer mode
that when activated, it avoids the synchronization
initialization and by doing this doesn't deletes other
developers networks.